Extract DropDownButton
