+++ site/src/components/DropdownButton/DropdownContent/DropdownContent.tsx
@@ -1,24 +1,21 @@
 import { makeStyles } from "@material-ui/core/styles"
-import { FC } from "react"
-import { ButtonMapping, ButtonTypesEnum } from "../constants"
+import { FC, ReactNode } from "react"
 
 export interface DropdownContentProps {
-  secondaryActions: ButtonTypesEnum[]
-  buttonMapping: Partial<ButtonMapping>
+  secondaryActions: Array<{ action: string, button: ReactNode }>
 }
 
 /* secondary workspace CTAs */
 export const DropdownContent: FC<React.PropsWithChildren<DropdownContentProps>> = ({
   secondaryActions,
-  buttonMapping,
 }) => {
   const styles = useStyles()
 
   return (
     <span data-testid="secondary-ctas">
-      {secondaryActions.map((action) => (
+      {secondaryActions.map(({ action, button }) => (
         <div key={action} className={styles.popoverActionButton}>
-          {buttonMapping[action]}
+          {button}
         </div>
       ))}
     </span>
